fix several concurrency-related issues
1. define GC_THREAD and incldue gc.h after pthread.h even if gc is not
used

2. Change the type of obj_size in classInfo to int64_t. LLVM emits int64
while our runtime used int32 before. It led to a bug that isInterface
bool was not set correctly.

3. Supply an empty checkExceptions array when construting
DeclaredMethods to avoid null pointer exceptions.

+++ runtime/native/class.h
@@ -64,7 +64,7 @@ struct JavaClassInfo {
     char *name;
     jclass *super_ptr;
     void *cdv;        // is a DispatchVector*
-    int32_t obj_size; // for array, it does not include data_size
+    int64_t obj_size; // for array, it does not include data_size
 
     jboolean isIntf;
